oracle-ebs-gemini-enterprise-framework/Agents/requirements.txt
# Union of all sub-agent requirements plus master-agent dependencies.
# Keep pinned versions in sync with sub-agent requirements.txt files.
google-adk==1.31.1
google-cloud-aiplatform==1.148.1
google-auth==2.49.2
mcp==1.27.0
matplotlib==3.10.8
opentelemetry-instrumentation-google-genai==0.7b0
opentelemetry-instrumentation-httpx==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-instrumentation-grpc==0.62b0